+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Upcoming
 
+## 1.1.5
+
 - fix: remove invalid sync cache mappings when Rekordbox track was deleted
 - fix: exit search after initial query instead of entering interactive mode